Add files via upload
My first PHP project on Github. Actually it is my second project: my first project was a login-system programmed the procedural way. After more or less finishing that project I changed it to an object oriented programmed (OOP) login-system. Which I decided to be my first portfolio project on Github. Note that the comments in the code are in Dutch.